# phase-3/specs/003-todo-ai-chatbot/contracts/openapi.yaml
openapi: "3.1.0"
info:
  title: Todo AI Chatbot API (Modern UI Edition)
  version: "2.0.0"
  description: |
    AI-powered todo management chatbot API with full CRUD,
    authentication, statistics, and chat endpoints.

servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://todo-ai-chatbot.railway.app
    description: Production (Railway)

paths:
  /api/{user_id}/chat:
    post:
      operationId: chat
      summary: Send a chat message to the AI todo assistant
      description: |
        Accepts a natural language message, processes it through
        the AI agent with MCP tools, and returns the assistant's
        response along with any tool calls that were executed.
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            minLength: 1
          description: User identifier for task/conversation scoping
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
      responses:
        "200":
          description: Successful chat response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
        "400":
          description: Bad request (empty message, message too long)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Validation error (malformed request body)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "429":
          description: Rate limit exceeded (100 req/min/user)
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds to wait before retrying
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Service temporarily unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/{user_id}/tasks:
    get:
      operationId: listTasks
      summary: List user tasks with optional filters
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [all, pending, completed]
            default: all
          description: Filter by task status
        - name: search
          in: query
          required: false
          schema:
            type: string
          description: Search tasks by title (case-insensitive)
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: object
                required: [tasks]
                properties:
                  tasks:
                    type: array
                    items:
                      $ref: "#/components/schemas/TaskResponse"
    post:
      operationId: createTask
      summary: Create a new task directly (non-chat)
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"

  /api/{user_id}/tasks/{task_id}:
    patch:
      operationId: updateTask
      summary: Update a task (title, description, completed)
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTask
      summary: Delete a task permanently
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Task deleted
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/{user_id}/stats:
    get:
      operationId: getUserStats
      summary: Get task statistics and activity data
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsResponse"

  /api/{user_id}/conversations:
    get:
      operationId: listConversations
      summary: List user conversations
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: object
                required: [conversations]
                properties:
                  conversations:
                    type: array
                    items:
                      $ref: "#/components/schemas/ConversationResponse"

  /api/{user_id}/conversations/{conversation_id}/messages:
    get:
      operationId: getConversationMessages
      summary: Get messages for a conversation
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: conversation_id
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
            maximum: 100
      responses:
        "200":
          description: Conversation messages
          content:
            application/json:
              schema:
                type: object
                required: [messages]
                properties:
                  messages:
                    type: array
                    items:
                      $ref: "#/components/schemas/MessageResponse"

  /api/{user_id}/profile:
    get:
      operationId: getProfile
      summary: Get user profile
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileResponse"
    patch:
      operationId: updateProfile
      summary: Update user profile (display name, theme)
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileResponse"

  /health:
    get:
      operationId: healthCheck
      summary: Health check endpoint
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: ["ok"]
                required:
                  - status

components:
  schemas:
    ChatRequest:
      type: object
      required:
        - message
      properties:
        conversation_id:
          type: integer
          nullable: true
          description: Existing conversation ID to continue
        message:
          type: string
          minLength: 1
          maxLength: 10000
          description: Natural language message from the user

    ChatResponse:
      type: object
      required:
        - conversation_id
        - response
        - tool_calls
      properties:
        conversation_id:
          type: integer
        response:
          type: string
        tool_calls:
          type: array
          items:
            $ref: "#/components/schemas/ToolCall"

    ToolCall:
      type: object
      required:
        - tool
        - parameters
        - result
      properties:
        tool:
          type: string
          enum:
            - add_task
            - list_tasks
            - complete_task
            - delete_task
            - update_task
        parameters:
          type: object
          additionalProperties: true
        result:
          type: object
          additionalProperties: true

    TaskResponse:
      type: object
      required:
        - id
        - title
        - completed
        - created_at
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        completed:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    CreateTaskRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          nullable: true

    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          nullable: true
        completed:
          type: boolean

    StatsResponse:
      type: object
      required:
        - total
        - pending
        - completed
        - recent_tasks
        - activity
      properties:
        total:
          type: integer
        pending:
          type: integer
        completed:
          type: integer
        recent_tasks:
          type: array
          maxItems: 5
          items:
            $ref: "#/components/schemas/TaskResponse"
        activity:
          type: array
          description: Daily task completion counts for last 7 days
          items:
            type: object
            required:
              - date
              - count
            properties:
              date:
                type: string
                format: date
              count:
                type: integer

    ConversationResponse:
      type: object
      required:
        - id
        - created_at
        - message_count
      properties:
        id:
          type: integer
        created_at:
          type: string
          format: date-time
        message_count:
          type: integer
        last_message_preview:
          type: string
          nullable: true

    MessageResponse:
      type: object
      required:
        - id
        - role
        - content
        - created_at
      properties:
        id:
          type: integer
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        tool_calls:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/ToolCall"
        created_at:
          type: string
          format: date-time

    ProfileResponse:
      type: object
      required:
        - id
        - email
      properties:
        id:
          type: string
        email:
          type: string
        display_name:
          type: string
          nullable: true
        theme_preference:
          type: string
          nullable: true
          enum: [light, dark, null]

    UpdateProfileRequest:
      type: object
      properties:
        display_name:
          type: string
          maxLength: 100
        theme_preference:
          type: string
          nullable: true
          enum: [light, dark, null]

    ErrorResponse:
      type: object
      required:
        - error
        - detail
      properties:
        error:
          type: string
        detail:
          type: string

    ValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  type: string
              msg:
                type: string
              type:
                type: string